pyproject.toml
[tool.poetry]
name = "chaosrings"
version = "1.2"
description = "A simple app that superimposes an image on top of another, centered."
authors = ["Stano"]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.23"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
